+++ src/Castle.Facilities.Quartz/IReleasingJobListener.cs
@@ -1,0 +1,8 @@
+﻿using Quartz;
+
+namespace Castle.Facilities.Quartz
+{
+    public interface IReleasingJobListener : IJobListener
+    {
+    }
+}